Implemented ComicsFragment. Created more observables to fetch Comics Data.

+++ app/src/main/java/com/kevin/marvellookup/pojo/ComicsPOJO/Result.java
@@ -11,7 +11,7 @@ public class Result {
     private Integer id;
     private Integer digitalId;
     private String title;
-    private Integer issueNumber;
+    private Double issueNumber;
     private String variantDescription;
     private String description;
     private String modified;
@@ -98,7 +98,7 @@ public class Result {
      * @return
      *     The issueNumber
      */
-    public Integer getIssueNumber() {
+    public Double getIssueNumber() {
         return issueNumber;
     }
 
@@ -107,7 +107,7 @@ public class Result {
      * @param issueNumber
      *     The issueNumber
      */
-    public void setIssueNumber(Integer issueNumber) {
+    public void setIssueNumber(Double issueNumber) {
         this.issueNumber = issueNumber;
     }
